% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCI.R
\encoding{UTF-8}
\name{UCI}
\alias{UCI}
\title{UCI: U-statistic permutation test of conditional independence}
\usage{
UCI(dat = dat, l1 = 0, l2 = 0, B = 199, weight = TRUE)
}
\arguments{
\item{dat}{a three-way frequency table}

\item{l1}{the dimension of \eqn{X}}

\item{l2}{the dimension of \eqn{Y}}

\item{B}{the number of random permutations}

\item{weight}{a logical indicating whether to consider weights; If \code{weight} is \code{TRUE}, the function uses
\eqn{T_W^\dagger} based on weighted U-statistics as a test statistic. Otherwise, the function uses \eqn{T} based on unweighted U-statistics
as a test statistic. See \insertCite{kim2022;textual}{UCI} for details.}
}
\value{
Returns the p-value for \code{UCI}-test and the value of the test statistic
}
\description{
Given a three-way contingency table, the function \code{UCI} computes two test statistics for conditional independence
depending on \code{weight} option. When \code{weight} is \code{TRUE}, it calculates the test statistic \eqn{T_{W,\mathrm{plug}}}
based on weighted U-statistics proposed by \insertCite{kim2022;textual}{UCI}.
On the other hand, when \code{weight} is \code{FALSE}, it calculates the test statistic \eqn{T} based on
unweighted U-statistics proposed by \insertCite{canonne2018testing;textual}{UCI}. See \insertCite{kim2022;textual}{UCI} for more details.
Both test statistics are calibrated by the Monte Carlo permutation method, yielding valid p-values.
Note that the computation of \eqn{T_{W,\mathrm{plug}}} requires the knowledge of \code{l1} (the dimension of \eqn{X}) and
\code{l2} (the dimension of \eqn{Y}). When these are unknown, then the function estimates \code{l1}
by the length of the unique elements of \eqn{X} in the dataset (and \code{l2} is similarly estimated).
}
\examples{
## sample size and dimensions
n = 100; l1 = 5; l2 = 5; d = 5

## Generate categorical data
x = sample(1:l1, n, replace = TRUE)
y = sample(1:l2, n, replace = TRUE)
z = sample(1:d, n, replace = TRUE)

## Generate the corresponding three-way frequency table
freq = table(x, y, z)

## Implement the test
# weighted statistic
UCI(freq, l1, l2, 199, TRUE)
# unweighted statistic
UCI(freq, 199, FALSE)
}
\references{
{\insertAllCited{}}
}
\seealso{
\code{\link{chi2test}} and \code{\link{gtest}}
}
