% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci.R, R/uci_package.R
\docType{package}
\name{uci}
\alias{uci}
\alias{_PACKAGE}
\alias{uci-package}
\title{Urban Centrality Index}
\usage{
uci(sf_object, var_name, full_border = FALSE)
}
\arguments{
\item{sf_object}{A \code{sf}. A number indicating the travel cost cutoff.}

\item{var_name}{A \code{string}. A col name}

\item{full_border}{A \code{lo}. t}
}
\description{
Urban Centrality Index

Calculate Urban Centrality Index
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://rafapereirabr.github.io/uci/}{website}.
}

\examples{
# load data
data_dir <- system.file("extdata", package = "uci")
grid <- readRDS(file.path(data_dir, "grid_bho.rds"))

# calculate UCI
 df <- uci(
        sf_object = grid,
        var_name = 'jobs',
        full_border = TRUE
        )
head(df)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rafapereirabr/uci}
  \item Report bugs at \url{https://github.com/rafapereirabr/uci/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

}
\concept{urban centrality index}
