% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci.R
\name{uci}
\alias{uci}
\title{Cumulative access based on a travel cost cutoff}
\usage{
uci(sf_object, var_name, full_border = FALSE)
}
\arguments{
\item{sf_object}{A \code{sf}. A number indicating the travel cost cutoff.}

\item{var_name}{A \code{string}. A col name}

\item{full_border}{A \code{lo}. t}
}
\description{
Calculates the number of opportunities accessible under a given specified
travel cost cutoff.
}
\examples{
data_dir <- system.file("extdata", package = "uci")
grid <- readRDS(file.path(data_dir, "grid_bho.rds"))

# active accessibility: number of schools accessible from each origin
 df <- uci(
        sf_object = grid,
        var_name = 'jobs',
        full_border = TRUE
        )
head(df)

}
\concept{urban centrality index}
