% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci.R, R/uci_package.R
\docType{package}
\name{uci}
\alias{uci}
\alias{_PACKAGE}
\alias{uci-package}
\title{Urban Centrality Index}
\usage{
uci(sf_object, var_name, full_border = FALSE)
}
\arguments{
\item{sf_object}{A \verb{POLYGON sf data.frame} of the study area.}

\item{var_name}{A \code{string}. The name of the column in \code{sf_object} with the
number of activities/opportunities/resources/services to be considered
when calculating urban centrality levels.}

\item{full_border}{A \code{lo}. t}
}
\description{
Calculates the Urban Centrality Index (UCI) as in Pereira et al., (2013) \url{https://doi.org/10.1111/gean.12002}. The UCI measures the
extent to which the spatial organization of a city or region varies from
extreme monocentric to extreme polycentric in a continuous scale from 0 to 1.
Values close to 0 indicate more polycentric patterns and values close to 1
indicate a more monocentric urban form.

Calculate Urban Centrality Index
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://rafapereirabr.github.io/uci/}{website}.
}

\examples{
# load data
data_dir <- system.file("extdata", package = "uci")
grid <- readRDS(file.path(data_dir, "grid_bho.rds"))

# calculate UCI
 df <- uci(
        sf_object = grid,
        var_name = 'jobs',
        full_border = TRUE
        )
head(df)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rafapereirabr/uci}
  \item Report bugs at \url{https://github.com/rafapereirabr/uci/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

}
\concept{urban centrality index}
