% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci.R, R/uci_package.R
\docType{package}
\name{uci}
\alias{uci}
\alias{_PACKAGE}
\alias{uci-package}
\title{Urban Centrality Index}
\usage{
uci(sf_object, var_name, boostrap_border = FALSE, showProgress = TRUE)
}
\arguments{
\item{sf_object}{A \verb{POLYGON sf data.frame} of the study area.}

\item{var_name}{A \code{string}. The name of the column in \code{sf_object} with the
number of activities/opportunities/resources/services to be considered
when calculating urban centrality levels.}

\item{boostrap_border}{A \code{logical}. The calculation of UCI requires one to
find the maximum value of the Venables spatial separation index of the
study area. If \code{boostrap_border = FALSE} (Default), the function uses
a heuristic approach that assumes that the max spatial separation would
occur when all activities were equaly distributed along the border of
the study  area. This is a fast approach, but it does not reach the
maximum spatial separation. Alternatively, if \code{boostrap_border = FALSE},
the function uses a boostrap approach that simulates 20000 random
distributions of activities along the border and uses the max spatial
separation found. This approach is more computationally expensive and
although it might not return the maximum theoretical value of spatial
separation, it is probably very close to it.}

\item{showProgress}{A \code{logical}. Indicates whether to show a progress bar for the
boostrap simulation. Defaults to \code{TRUE}.}
}
\description{
Calculates the Urban Centrality Index (UCI) as in Pereira et al., (2013) \url{https://doi.org/10.1111/gean.12002}.
The UCI measures the extent to which the spatial organization of a city or
region varies from extreme monocentric to extreme polycentric in a continuous
scale from 0 to 1. Values close to 0 indicate more polycentric patterns and
values close to 1 indicate a more monocentric urban form.

Calculate Urban Centrality Index
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://rafapereirabr.github.io/uci/}{website}.
}

\examples{
# load data
data_dir <- system.file("extdata", package = "uci")
grid <- readRDS(file.path(data_dir, "grid_bho.rds"))

# calculate UCI
df <- uci(
        sf_object = grid,
        var_name = 'jobs',
        boostrap_border = FALSE
        )
head(df)

# calculate UCI with boostrap
df2 <- uci(
        sf_object = grid,
        var_name = 'jobs',
        boostrap_border = TRUE,
        showProgress = TRUE
        )
head(df2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rafapereirabr/uci}
  \item Report bugs at \url{https://github.com/rafapereirabr/uci/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

}
\concept{urban centrality index}
